fix: 타워 사거리 소수점 적용

+++ Programs/GT9-Team3/Assets/Tower/Scripts/Loader/TowerDataRow.cs
@@ -54,7 +54,7 @@ public class TowerDataRow
     /// <summary>
     /// Attack_Range
     /// </summary>
-    public int Attack_Range;
+    public float Attack_Range;
 
     /// <summary>
     /// Attack_Speed
